+++ changelog.md
@@ -1,5 +1,13 @@
 # sub-domain-scanner changelog
 
+## v1.3.0
+* Add CLI method discover-domains
+* Add auto wildcard output and dedupe output for discover-domains
+* Dedupe output from discover-hostnames
+* Add wildcard subsitution for common sub-domains for discover-hostnames output
+* Minor fixes and tidies
+* Try to fix Travis (this pkg requires node 11 currently)
+
 ## v1.2.0
 * Add JSON output format for test-hostnames
 * Output all hostnames in test-hostnames